Module Graph_gtk.DGraphView

View classes.

Each optional function delay_node, delay_edge and delay_cluster of this module may be used to indicate whether an element must be displayed instantaneously (if the function returns false) or may be delayed for latter display (if the function returns true). By default, each function always returns false. It may be set for returning true from time to time, improving efficiency.

class type ['vertex, 'edge, 'cluster] view = object ... end

Graph widget derived from GnoCanvas.canvas. Support zooming and scrolling.

module type S = sig ... end

module Make (V : Graph.Sig.HASHABLE) (E : Graph.Sig.HASHABLE) (C : Graph.Sig.HASHABLE) : S with type vertex = V.t and type edge = E.t and type cluster = C.t