Class type Graph_gtk__DGraphView.view

Graph widget derived from GnoCanvas.canvas. Support zooming and scrolling.

inherit GnoCanvas.canvas

method model : ('vertex, 'edge, 'cluster) Graph.DGraphModel.abstract_model

Getters

method get_node : 'vertex -> 'vertex Graph_gtk.DGraphViewItem.view_item

method get_edge : 'edge -> 'edge Graph_gtk.DGraphViewItem.view_item

method get_cluster : 'cluster -> 'cluster Graph_gtk.DGraphViewItem.view_item

Iterators

method iter_nodes : ('vertex Graph_gtk.DGraphViewItem.view_item -> unit) -> unit

method iter_edges : ('vertex Graph_gtk.DGraphViewItem.view_item -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit) -> unit

method iter_edges_e : ('edge Graph_gtk.DGraphViewItem.view_item -> unit) -> unit

method iter_clusters : ('cluster Graph_gtk.DGraphViewItem.view_item -> unit) -> unit

method iter_succ : ('vertex Graph_gtk.DGraphViewItem.view_item -> unit) -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

method iter_pred : ('vertex Graph_gtk.DGraphViewItem.view_item -> unit) -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

method iter_succ_e : ('edge Graph_gtk.DGraphViewItem.view_item -> unit) -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

method iter_pred_e : ('edge Graph_gtk.DGraphViewItem.view_item -> unit) -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

method iter_associated_vertex : ('vertex Graph_gtk.DGraphViewItem.view_item -> unit) -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

Membership functions

method mem_edge : 'vertex Graph_gtk.DGraphViewItem.view_item -> 'vertex Graph_gtk.DGraphViewItem.view_item -> bool

method find_edge : 'vertex Graph_gtk.DGraphViewItem.view_item -> 'vertex Graph_gtk.DGraphViewItem.view_item -> 'edge Graph_gtk.DGraphViewItem.view_item

method src : 'edge Graph_gtk.DGraphViewItem.view_item -> 'vertex Graph_gtk.DGraphViewItem.view_item

method dst : 'edge Graph_gtk.DGraphViewItem.view_item -> 'vertex Graph_gtk.DGraphViewItem.view_item

Zooming

method zoom_factor : float

The current zoom factor.

method zoom_to : float -> unit

Set an absolute zoom factor.

method zoom_in : unit -> unit

Increase zoom_factor by zoom_factor*zoom_padding.

method zoom_out : unit -> unit

Decrease zoom_factor by zoom_factor*zoom_padding.

method adapt_zoom : unit -> unit

Zoom in order to view the whole graph (bird eye view).

method set_zoom_padding : float -> unit

Set the zoom padding used by zoom_in and zoom_out. It defaults to 0.1.

method center_node : 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

Center canvas on a node.

Highlighting

method connect_highlighting_event : unit -> unit

method highlight : ?⁠color:(int32 * int32) -> 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

Change the color of the given vertex item. May be cancelled by dehighlight. If color is primary,secondary, then primary is used except if the current color is primary. In this case, secondary is used.

method dehighlight : 'vertex Graph_gtk.DGraphViewItem.view_item -> unit

Cancel highlight.